Refactor add experiment to dataset and ensure no unprocessed samples are added
