+++ src/state/download/DataSetManager.js
@@ -47,7 +47,7 @@ export default class DataSetManager {
         return sampleAccessions.indexOf(sample) === -1;
       });
       
-      if (filteredSamples.length) {
+      if (filteredSamples.length > 0) {
         result[accessionCode] = filteredSamples;
       }
       return result;
@@ -56,5 +56,38 @@ export default class DataSetManager {
     return newDataSet;
   }
 
+  /**
+   * Adds a set of experiments to the current dataset
+   * @param {array<{accession_code, samples}>} experiments 
+   */
+  addExperiment(experiments) {
+    let newDataSetExperiments = {};
+    for (let experiment of experiments) {
+      if (experiment.samples.length === 0) continue;
+      let sampleAccessions = experiment.samples
+        .filter(x => x.is_processed)
+        .map(x => x.accession_code);
+      newDataSetExperiments[experiment.accession_code] = sampleAccessions;
 
+      // check if the current experiment had samples in the dataset previousle
+      // in which case make sure that those are also added
+      if (this.dataSet[experiment.accession_code]) {
+        // Remove duplicates from the array, since the backend throws errors
+        // on non-unique accessions
+        newDataSetExperiments[experiment.accession_code] = [
+          ...new Set([
+            ...this.dataSet[experiment.accession_code],
+            ...sampleAccessions
+          ])
+        ];
+      }
+    }
+
+    return {
+      ...this.dataSet,
+      ...newDataSetExperiments
+    };
+  }
 }
+
+
